Do not use portal deployment key
Make use of profile_public_key instead of portal_public_key_path
Not knowing in advance the name of the default user of the OS, requires to ask
to the user this information. In order to avoid to many settings the deployment
will not use the portal deployment key (and then switch in the Ansible step),
but just inject the public key of the user.